Split script into multiple python files based on functionality
